Embed argument parsing as part of plugin definition
